CHG: Memory problem 4
Solutions 2 and 3 were either red herrings or weren't enough to solve the problem - the view controllers aren't being destroyed, because the timer is creating a strong refeence cycle.

There are several solutions here: rewrite the code using smarter animations, use a weak self closure capture list, or destroy the timer when it's no longer needed, thus breaking the cycle.
We take the last option and stop the timer in viewWillDisappear()